requirements.txt
langchain
openai
streamlit
google-auth
google-auth-httplib2
google-auth-oauthlib
google-api-python-client
oauthlib
python-dotenv
colorama
PyGithub
googleapiclient
